requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==1.2.2
Pillow==9.0.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
tzdata==2021.5
